+++ packages/breadcrumbs/CHANGELOG.md
@@ -35,6 +35,7 @@
 ### v2.1.0
 
 - Support react router
+- Change npm run watch browser-sync location
 
 
 ### v2.0.3